rerewriting of API
